refactor nshm package to nshmp
