+++ src/gov/usgs/earthquake/nshmp/www/AdminService.java
@@ -1,4 +1,4 @@
-package gov.usgs.earthquake.nshm.www;
+package gov.usgs.earthquake.nshmp.www;
 
 import java.io.IOException;
