+++ src/gov/usgs/earthquake/nshmp/www/XY_DataGroup.java
@@ -1,4 +1,4 @@
-package gov.usgs.earthquake.nshm.www;
+package gov.usgs.earthquake.nshmp.www;
 
 import java.util.ArrayList;
 import java.util.List;
