+++ src/gov/usgs/earthquake/nshmp/www/meta/Constrained.java
@@ -1,4 +1,4 @@
-package gov.usgs.earthquake.nshm.www.meta;
+package gov.usgs.earthquake.nshmp.www.meta;
 
 /**
  * Interface implemented by enum parameters that impose restrictions on other
