+++ src/gov/usgs/earthquake/nshmp/www/meta/package-info.java
@@ -4,6 +4,6 @@
  * <p>Classes in this package largely provide support for web services used on
  * the public facing USGS website. Services that are not public facing may use a
  * simpler metadata structure defined within the service class itself (e.g.
- * {@link gov.usgs.earthquake.nshm.www.SpectraService}.
+ * {@link gov.usgs.earthquake.nshmp.www.SpectraService}.
  */
-package gov.usgs.earthquake.nshm.www.meta;
+package gov.usgs.earthquake.nshmp.www.meta;
